Move RecordCamera2Fragment to fragment package
