+++ litr-demo/src/main/java/com/linkedin/android/litr/demo/fragment/CameraSizes.kt
@@ -7,7 +7,7 @@
  *
  * Author: Ian Bird
  */
-package com.linkedin.android.litr.demo
+package com.linkedin.android.litr.demo.fragment
 
 import android.graphics.Point
 import android.hardware.camera2.CameraCharacteristics
